Bot arbitrage/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
aiohttp==3.9.1
asyncio==3.4.3
pydantic==2.5.0
typing-extensions==4.8.0
lighter-python-sdk
